requirements.txt
aiofiles==0.6.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.17.0
animegifs==0.7.2
async-generator==1.10
async-timeout==4.0.3
asyncio-extras==1.3.2
asyncpraw==7.7.1
asyncprawcore==2.3.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
cleverbotfreeapi==1.1.1
colorama==0.4.6
decorator==4.4.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.8
mal-api==0.5.3
moviepy==1.0.3
multidict==6.0.4
numpy==1.25.2
Pillow==10.0.0
proglog==0.1.10
PyJWT==2.8.0
pytube==15.0.0
RedDownloader==4.2.0
redvid==2.0.2
requests==2.31.0
soupsieve==2.5
tqdm==4.62.3
typing_extensions==4.7.1
update-checker==0.18.0
urllib3==2.0.4
yarl==1.9.2
